Added some Biginteger utility functions. Fixed some typing problems with the RTE Packet.

+++ app/src/main/java/rte/RTEFrame.kt
@@ -1,9 +1,10 @@
 package rte
 
 import android.graphics.Bitmap
+import java.math.BigInteger
 
 /**
  * Created by jk on 2/26/18.
  * Just a filler class for now, may be necessary for future use by adding timestamps.
  */
-data class RTEFrame(var bitmap: Bitmap, var fid: Int, var timestamp: Long)
+data class RTEFrame(var bitmap: Bitmap, var fid: Long, var timestamp: BigInteger)
